Search Activity implemented
The file build.gradle was updated; make sure to update that. SearchActivity is a simple activity in the front end that has a search bar and a button to search. When the user presses "send", searchActivity starts the PageListActivity and sends it the string in the searchView; this was tested with a Log.cat. The components dummyContent and toolbar were giving me bugs, so I commented them out. Need to figure out where to instantiate the dbOperations and how to run the query.