Interface: UI activities can open DB and make query
PageListActivity instantiates the DB. SearchActivity now only passes the keyword to the PageListActivity. When user enters the keyword in the SearchView of PageListActivity, query is made. Results are only shown in the Log right now. Need to update PageListActivity and PageDetailsActivity to reflect query results. Currently, query does not search through the titles of articles. A hard-coded article with title "title" and content "content' was introduced. Searching for "title" does not return anything, but searching for "content" returns the correct article.