Prettify step 1: Remove [[Image]] from articles
To prettify the PHArticles' contents, parse through content and edit as needed. This takes a long time. If we try to do this in the method parsePagesAndPopulateDB(), then the DB will get garbage data, because it takes too long to get the JSON and parse through all the pages. All the PHArticles made in parsePagesAndPopulateDB() are stored in a temporary global ArrayList and then cleaned of unwanted stuff. After that, add to database normally. This iteration only searches for [[Image]] strings that normally take the place of jpg's. There is more to do