Spell check implemented
When the database is populated, we insert every word from every article title into a table that stores unique words from titles, and make sure that each value in the table is unique. After populating is done, read the values from the table of title words into a global array list. Then, when the user enters a search word, the app computes the Levenshtein between the given search word and every word in the global array list of title words and searches using the word that is closest to an existing title word