Arts database download is stable
CategorySelectionActivity downloads the database from the server and finds the path to the downloaded file (which was a bitch to find all the code. I swear Android made this difficult on purpose). CategorySelectionActivity also copies everything from the downloaded database into the PocketHow.db, because the downloaded database file path was gibberish when I passed it to the searchActivity. CategorySelectionActivity does not call the makeRequests() function, and it does not add anything to the Intent when starting the searchActivity, so it won't fetch or parse articles from WikiHow. The code in searchActivity that retrieves the Bundle from the Intent will throw and catch an error, so it does not call populateDB(). DbOperations doesn't search for the closest word if the search word is null or just spaces.