Add PHCategory to category_to_pageid table
We now make categories. There used to be an issue where pageIdList had a bunch of commas, so it interpreted each comma as a different parameter to the insert, so we needed to add quotation marks around it. The same problem happens for category strings with spaces. SQL hates spaces, so we need to put quotation marks around it