Remove category icons when user downloads them
When the user selects the categories to be downloaded, the gridview removes the icons before downloading them to prevent the user from jumping back from searchActivity and interrupting the download process. If the user jumps back to CategorySelectionActivity before the database is done downloading, then the user cannot redownload the database because it cannot be accessed via gridview, and the app will continue downloading and loading as  normal. If something goes wrong, then the app will put the category icon back on the gridview